Switched map tile storage to HashMap<TilePosition, Tile> from ArrayList<Tile>, moved files into new package
