--- src/render/GuiMenuListener.java
@@ -1,11 +1,0 @@
-package render;
-
-public interface GuiMenuListener {
-	/**
-	 * Listens for when an option in a menu is selected, key is the text/title
-	 * of the button pressed
-	 * 
-	 * @param key
-	 */
-	public void buttonSelected(String key);
-}
