--- src/scenes/Scene.java
@@ -1,23 +1,0 @@
-package scenes;
-
-import java.awt.Graphics2D;
-
-/**
- * Contains basic container methods for scenes, also the SceneManager handle for
- * each Scene
- * 
- * @author Robert
- *
- */
-public abstract class Scene {
-
-	protected SceneManager mgr;
-
-	public Scene(SceneManager mgr) {
-		this.mgr = mgr;
-	}
-
-	public abstract void update();
-
-	public abstract void render(Graphics2D g);
-}
